Регистрация через RegistrationService, дабы избежать цикла.
