+++ src/main/java/ru/division/of/expenses/app/configs/JwtRequestFilter.java
@@ -21,7 +21,10 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @Slf4j
 public class JwtRequestFilter extends OncePerRequestFilter {
-//    private final UserDetailsService userDetailsService;
+    ///////////////  Если делать регистацию в UserService,
+    // то это поле нужно закоммитить, дабы избежать цикла.
+    private final UserDetailsService userDetailsService;
+    ////////////////
     private final JwtTokenUtil jwtTokenUtil;
 
     @Override
